services/backend/core/src/requirements.txt
click==8.1.3
colorama==0.4.6
cx-Oracle==8.3.0
DateTime==5.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==2.6.0
python-dotenv==0.21.0
requests==2.28.2
pytz==2022.7.1
shortuuid==1.0.11
six==1.16.0
SQLAlchemy==1.4.46
Werkzeug==2.2.2
zipp==3.11.0
zope.interface==5.5.2
#for obtainip.py
boto3==1.26.68
twilio==7.17.0
schedule==1.1.0
